test(desktop): keep runtime regression within lint budgets
Problem:
The project-tree regression coverage made two already-large desktop files exceed repository lint ratchets.

Root cause:
The cross-workspace test lived in topic_activation_test.go, and the runtime event comment exceeded the allowed local essay budget.

Fix:
Move the unchanged regression into a dedicated test file and compact the explanatory comment without raising any baseline.

Verification:
- go run ./tools/repolint
- go test ./... in desktop
- focused desktop go test -race
- go vet ./... in desktop
- project-tree frontend tests, typecheck, and hooks lint
- git diff --check
